Unit tests throug setup.py test, with backend in DEBUG mode and re-uses existing.
Suppressed stdout/stderr in front end unit tests.

+++ setup.py
@@ -14,17 +14,25 @@ with open('./README.md') as fhandle:
     readme = fhandle.read()
 
 _root_dir = os.path.dirname(os.path.realpath(__file__))
+cmake_dir = os.path.join(_root_dir, 'cmake-build-debug') # cmake target directory
 
 
-def _build_backend(root_dir):
+def _build_backend(root_dir, mode = "install"):
+    assert(mode in {"install", "test"})
+    build_type = "REL_WITH_ASSERTS" # Has -O flag
+    if mode == "test":
+        build_type = "DEBUG" # No optimisation: allows for easy debugging
+
     print('Compiling gramtools backend')
-    cmake_dir = os.path.join(root_dir, 'cmake-build-debug')
-    try:
-        shutil.rmtree(cmake_dir, ignore_errors=False)
-    except FileNotFoundError:
-        pass
-    os.mkdir(cmake_dir)
-    subprocess.call('CC=gcc CXX=g++ cmake -DCMAKE_BUILD_TYPE=REL_WITH_ASSERTS ..', cwd=cmake_dir, shell=True)
+
+    if not os.path.exists(cmake_dir):
+        os.mkdir(cmake_dir)
+    else:
+        if mode == "install": # Erasing pre-existing compiled source tree
+            shutil.rmtree(cmake_dir, ignore_errors=False)
+            os.mkdir(cmake_dir)
+
+    subprocess.call(f'CC=gcc CXX=g++ cmake -j 4 -DCMAKE_BUILD_TYPE={build_type} ..', cwd=cmake_dir, shell=True)
 
     return_code = subprocess.call(['make'], cwd=cmake_dir)
     if return_code != 0:
@@ -33,7 +41,11 @@ def _build_backend(root_dir):
 
 
 def _test_backend(root_dir):
-    cmake_dir = os.path.join(root_dir, 'cmake-build-debug')
+    print("""
+    ##########################
+    ##Running back end tests##
+    ##########################
+    """)
     test_runner = os.path.join(cmake_dir, 'libgramtools', 'submods', 'tests', 'test_main')
     test_dir = os.path.join(root_dir, 'libgramtools', 'submods', 'tests')
 
@@ -43,16 +55,22 @@ def _test_backend(root_dir):
         exit(-1)
 
 
-## Run the front end unit tests: operations making prg, `infer`, `quasimap`.
+## Run the front end unit tests: operations making prg, `infer`, integration tests.
 # Code adapted from https://stackoverflow.com/q/17001010
 def _test_frontend(root_dir):
-    print("Running gramtools front end unit tests")
+    print("""
+    ###########################
+    ##Running front end tests##
+    ###########################
+    """)
     # use the default shared TestLoader instance
     test_loader = unittest.defaultTestLoader
     test_suite = test_loader.discover(root_dir, pattern = "test*.py") # The pattern used is the default
 
     # use the basic test runner that outputs to sys.stderr
-    test_runner = unittest.TextTestRunner(stream = sys.stdout, verbosity = 2)
+    # `buffer` suppresses stdout/err if tests pass
+    test_runner = unittest.TextTestRunner(stream = sys.stdout, verbosity = 2,
+                                          buffer = True)
     tests_results = test_runner.run(test_suite)
 
     if not tests_results.wasSuccessful():
@@ -77,7 +95,7 @@ class _DevelopCommand(develop):
     Tests not ran either- these should be ran manually (eg python3 -m unittest from source root for front-end).
 
     Command to build develop: (do this inside development dir)
-        pip3 install -vvv --editable --process-dependency-links /path/to/gramtools  # wrapper round setuptools develop.
+        pip3 install -vvv --editable /path/to/gramtools  # wrapper round setuptools develop.
 
         Note: can also use `python3 /path/to/gramtools/setup.py develop` BUT will not properly run setup.py on packages.
 
@@ -92,16 +110,20 @@ class _TestCommand(test):
         python3 setup.py test
     """
     def run(self):
-        _build_backend(_root_dir)
+        # test.run(self) # Setuptools' own front end test logic
+        _test_frontend(_root_dir)
+        _build_backend(_root_dir, mode = "test")
         _test_backend(_root_dir)
-        test.run(self)
+        print("""
+        All tests passed.
+        """)
 
 
 _package_data = {
     'gramtools': [
         'bin/gram',
         'lib/*',
-        'utils/vcf_to_linear_prg.pl'],
+        'commands/build/vcf_to_linear_prg.pl'],
 }
